+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # CHANGELOG
 
+* `2.0.86` Release - [2.0.86](#2086)
 * `2.0.85` Release - [2.0.85](#2085)
 * `2.0.81` Release - [2.0.81](#2081)
 * `2.0.80` Release - [2.0.80](#2080)
@@ -7,6 +8,10 @@
 * `2.0.78` Release - [2.0.78](#2078)
 * `0.77.x` Releases - [0.77.0](#0770)
 
+## 2.0.86
+
+#### Bug Fixes
+* Use correct `FairplayRequester` when streaming *vod*
 
 ## 2.0.85